fix(modules.board): review fixes
