+++ packages/modules.board/src/shapes/geo/index.ts
@@ -2,3 +2,4 @@ export { XiGeoShapeUtil } from './XiGeoShapeUtil';
 export { BorderPicker } from './BorderPicker';
 export { FillTypePicker } from './FillTypePicker';
 export { XiGeoTool } from './XiGeoTool';
+export { XiGeoComponent } from './XiGeoComponent';
