+++ packages/modules.board/src/ui/components/popups/Pen/ColorSet.tsx
@@ -5,20 +5,6 @@ import { ColorDot } from '../../canvas';
 
 type ColorOptionT = (typeof colorOptions)[number]['name'];
 
-// const ColorCircle = ({ colorClass, isSelected, handleClick }: ColorCircleT) => (
-//   <div
-//     className={`m-auto flex h-[36px] w-[36px] items-center justify-center rounded-full ${isSelected ? 'border-grey-100 border' : 'border border-transparent'}`}
-//   >
-//     <button
-//       onClick={() => handleClick()}
-//       type="button"
-//       aria-label={`Select ${colorClass}`}
-//       tabIndex={0}
-//       className={`h-[31px] w-[31px] cursor-pointer rounded-full ${colorClass}`}
-//     />
-//   </div>
-// );
-
 type ColorGridProps = {
   currentColor?: string;
 };
